refactor(pipeline): move parallel-review integration tests to review_test.go [#288]
Move 5 parallel-review integration tests from engine_test.go to
review_test.go: ReviewReworkRouting, ReviewReworkFeedbackInjected,
SkippedReviewPhaseReworkSignalRoutesToImplement,
FollowUpPhase_RunsOnMinorFindings, and
ReworkFeedbackIncludesPriorReviewCycles.

Leave ReviewReworkDefaultMaxCycles, ReviewReworkCyclesPersisted, and
ReviewReworkAndPatchIndependent in engine_test.go since they test
engine config/state behavior rather than review.go functions.
